created seperate admin panel
